fix: build error

+++ source/MaaFramework/Vision/FeatureMatcher.cpp
@@ -190,7 +190,7 @@ FeatureMatcher::ResultsVec FeatureMatcher::feature_postproc(
     };
 
     ResultsVec results;
-    while (matches.size() >= param_.count) {
+    while (matches.size() >= static_cast<size_t>(param_.count)) {
         cv::Mat homography = cv::findHomography(obj, scene, cv::RANSAC);
         if (homography.empty()) {
             break;
